requirements.txt
openai==2.6.1
fastapi==0.120.0
pydantic==2.12.3
uvicorn[standard]==0.38.0
python-dotenv==1.1.1
ipykernel
pandas==2.2.3
tqdm==4.67.1
# trl==0.22.2
# transformers==4.56.2
modal
httpx
streamlit
requests